More fixups and cleanups to Player/Stream/Output interfaces.

+++ src/square/ConsoleUI.cpp
@@ -271,15 +271,15 @@ void ConsoleUI::SetVolume(Args args)
 {
     if (args.size() > 0)
     {
-        short newVolume = 0;
-        if (convertString<short>(newVolume, args[0]))
+        float newVolume = 0;
+        if (convertString<float>(newVolume, args[0]))
         {
             this->SetVolume(newVolume);
         }
     }
 }
 
-void ConsoleUI::SetVolume(short volume)
+void ConsoleUI::SetVolume(float volume)
 {
     transport.SetVolume(volume);
 }
